wrapped in a class
